[PhpBridge] use progressive enhancement to keep compatibility with PHP 5 in SymfonyTestsListener
