+++ src/Symfony/Bridge/PhpUnit/SymfonyTestsListener.php
@@ -11,9 +11,8 @@
 
 namespace Symfony\Bridge\PhpUnit;
 
+use PHPUnit\Framework\BaseTestListener;
 use PHPUnit\Framework\Test;
-use PHPUnit\Framework\TestListener;
-use PHPUnit\Framework\TestListenerDefaultImplementation;
 use PHPUnit\Framework\TestSuite;
 use PHPUnit\Framework\Warning;
 
@@ -21,8 +20,8 @@ use PHPUnit\Framework\Warning;
 // gets defined without executing the code before it and so the definition is not properly conditional)
 if (class_exists('PHPUnit_Runner_Version') && version_compare(\PHPUnit_Runner_Version::id(), '6.0.0', '<')) {
     class_alias('Symfony\Bridge\PhpUnit\Legacy\SymfonyTestsListener', 'Symfony\Bridge\PhpUnit\SymfonyTestsListener');
-} elseif (version_compare(\PHPUnit\Runner\Version::id(), '7.0.0', '<')) {
-    class_alias('Symfony\Bridge\PhpUnit\Legacy\SymfonyTestsListenerPhpunit6', 'Symfony\Bridge\PhpUnit\SymfonyTestsListener');
+} elseif (version_compare(\PHPUnit\Runner\Version::id(), '7.0.0', '>=')) {
+    class_alias('Symfony\Bridge\PhpUnit\SymfonyTestsListenerWithReturnTypes', 'Symfony\Bridge\PhpUnit\SymfonyTestsListener');
 } else {
     /**
      * Collects and replays skipped tests.
@@ -31,10 +30,8 @@ if (class_exists('PHPUnit_Runner_Version') && version_compare(\PHPUnit_Runner_Ve
      *
      * @final
      */
-    class SymfonyTestsListener implements TestListener
+    class SymfonyTestsListener extends BaseTestListener
     {
-        use TestListenerDefaultImplementation;
-
         private $trait;
 
         public function __construct(array $mockedNamespaces = array())
@@ -47,29 +44,29 @@ if (class_exists('PHPUnit_Runner_Version') && version_compare(\PHPUnit_Runner_Ve
             $this->trait->globalListenerDisabled();
         }
 
-        public function startTestSuite(TestSuite $suite): void
+        public function startTestSuite(TestSuite $suite)
         {
-            $this->trait->startTestSuite($suite);
+            return $this->trait->startTestSuite($suite);
         }
 
-        public function addSkippedTest(Test $test, \Throwable $t, float $time): void
+        public function addSkippedTest(Test $test, \Exception $e, $time)
         {
-            $this->trait->addSkippedTest($test, $t, $time);
+            return $this->trait->addSkippedTest($test, $e, $time);
         }
 
-        public function startTest(Test $test): void
+        public function startTest(Test $test)
         {
-            $this->trait->startTest($test);
+            return $this->trait->startTest($test);
         }
 
-        public function addWarning(Test $test, Warning $e, float $time): void
+        public function addWarning(Test $test, Warning $e, $time)
         {
-            $this->trait->addWarning($test, $e, $time);
+            return $this->trait->addWarning($test, $e, $time);
         }
 
-        public function endTest(Test $test, float $time): void
+        public function endTest(Test $test, $time)
         {
-            $this->trait->endTest($test, $time);
+            return $this->trait->endTest($test, $time);
         }
     }
 }
